# wacp-console/openapi.yaml
info:
  description: REST API for the WACP Console — coordination workbench for the WACP ecosystem.
  license:
    name: Apache-2.0
  title: WACP Console API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/auth/change-password:
    post:
      responses:
        '204':
          description: No content
        '401':
          description: Unauthenticated
      summary: Change password
      tags:
      - auth
...
  /api/auth/logout:
    post:
      responses:
        '204':
          description: No content
        '401':
          description: Unauthenticated
      summary: Logout
      tags:
      - auth
  /api/auth/whoami:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
      summary: Get current user
      tags:
      - auth
  /api/checkpoint-types:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List checkpoint types
      tags:
      - discovery
  /api/checkpoint-types/{name}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get checkpoint type
      tags:
      - discovery
  /api/envelope-types:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List envelope types
      tags:
      - discovery
  /api/envelope-types/{name}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get envelope type
      tags:
      - discovery
...
  /api/roles:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List roles
      tags:
      - discovery
  /api/roles/{id}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get role
      tags:
      - discovery
...
  /api/tools:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List tools
      tags:
      - discovery
  /api/tools/{name}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get tool
      tags:
      - discovery
...
  /api/verticals:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List verticals
      tags:
      - discovery
  /api/verticals/{id}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get vertical
      tags:
      - discovery
  /api/verticals/{id}/checkpoint-types:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get vertical checkpoint types
      tags:
      - verticals
...